requirements.txt
openai
streamlit
streamlit-chat
deta
streamlit-authenticator
python-dotenv